% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity_and_stability_checks.R
\name{array_size}
\alias{array_size}
\title{Check size of input measurements}
\usage{
array_size(x)
}
\arguments{
\item{x}{A numeric array}
}
\value{
sz the size of the array
}
\description{
Takes in any numeric array and prints out the size of the array.
}
