% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity_and_stability_checks.R
\name{get_files_from_folder}
\alias{get_files_from_folder}
\title{Get list of files in folder}
\usage{
get_files_from_folder(folder)
}
\arguments{
\item{folder}{Folder name}
}
\value{
flist List of files in folder
}
\description{
Takes in the folder name and gets all Excel files in it
}
