% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity_and_stability_checks.R
\name{get_mu}
\alias{get_mu}
\title{Get mu}
\usage{
get_mu(arr_in)
}
\arguments{
\item{arr_in}{An array of numbers}
}
\value{
mu Mu
}
\description{
Takes in an array of numbers and returns mu
}
