% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity_and_stability_checks.R
\name{get_mu_sigma}
\alias{get_mu_sigma}
\title{Get simple mu and sigma}
\usage{
get_mu_sigma(arr_in)
}
\arguments{
\item{arr_in}{An array of numbers}
}
\value{
arr_out An two elements array: mu and sigma
}
\description{
Takes in an array of numbers and returns mu and sigma
}
