% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity_and_stability_checks.R
\name{get_scatterplot}
\alias{get_scatterplot}
\title{Make scatterplot with errorbars}
\usage{
get_scatterplot(x_arrs, x_labs, x_u, fname, fldr)
}
\arguments{
\item{x_arrs}{An array of numbers}

\item{x_labs}{An array of x-labels}

\item{x_u}{An array of uncertainties}

\item{fldr}{Folder name where to save the chart}
}
\description{
Takes in an array of means, an array of labels and an array of uncertainties and makes scatterplot with errorbars
}
