% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity_and_stability_checks.R
\name{get_sigma}
\alias{get_sigma}
\title{Get sigma}
\usage{
get_sigma(arr_in)
}
\arguments{
\item{arr_in}{An array of numbers}
}
\value{
sigma Sigma
}
\description{
Takes in an array of numbers and returns sigma
}
