% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity_and_stability_checks.R
\name{make_bar_plot}
\alias{make_bar_plot}
\title{Makes bar plot}
\usage{
make_bar_plot(df, var_name)
}
\arguments{
\item{df}{frame}

\item{var_name}{var name as string}
}
\description{
takes in variable name and makes barplot
}
