% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity_and_stability_checks.R
\name{make_data_frame}
\alias{make_data_frame}
\title{Read xlsx files and store data}
\usage{
make_data_frame()
}
\value{
flist List of files in folder
}
\description{
Stores all data into appropriate R object
}
