% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity_and_stability_checks.R
\name{make_table}
\alias{make_table}
\title{takes in variable name and makes table}
\usage{
make_table(df, var_name)
}
\arguments{
\item{df}{frame}

\item{var_name}{var name as string}
}
\description{
takes in variable name and makes table
}
