% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity_and_stability_checks.R
\name{read_data}
\alias{read_data}
\title{Read xlsx files and store data}
\usage{
read_data()
}
\value{
flist List of files in folder
}
\description{
Stores all data into appropriate R object
}
