% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity_and_stability_checks.R
\name{z_score}
\alias{z_score}
\title{Get z-score}
\usage{
z_score(x, mu, sigma)
}
\arguments{
\item{x}{A number}

\item{mu}{Mean}

\item{sigma}{Sigma}
}
\value{
z The z-score
}
\description{
Takes in a single number, the mean and the sigma and calculates the z-score
}
